+++ app/src/main/java/de/th/ro/datavis/models/AtomicField.java
@@ -20,11 +20,11 @@ import de.th.ro.datavis.util.enums.InterpretationMode;
                 childColumns = "antennaId",
                 onDelete = CASCADE)})
 public class AtomicField {
-    public int getTilt() {
+    public double getTilt() {
         return tilt;
     }
 
-    public void setTilt(int tilt) {
+    public void setTilt(double tilt) {
         this.tilt = tilt;
     }
 
@@ -53,7 +53,7 @@ public class AtomicField {
         this.spheres = spheres;
     }
 
-    public int tilt;
+    public double tilt;
     public double frequency;
     @NonNull
     public InterpretationMode interpretationMode;
